Fixed bug # 1852032 where some app controller tests were leaving temp files on successful completion.
Also added new feature (and associated tests) to CommandLineAppController.getTmpFilename. A new parameter will add a class identifier derived from the class name, to the temp filenames, whcih will be appended following the prefix. This will allow us to easily determine what class created a temp file, which should be especially useful in case where extra files are left around after (e.g.) run_tests. I also deleted the CommandLineApplication static variables TmpPrefix and TmpSuffix. These were added on the sprint branch to address this issue, but it seems safer to just handle this on-the-fly b/c I'm guessing that specifying TmpPrefix is something that people would frequently forget to do. None of the built-in ApplicationControllers use TmpPrefix or TmpSuffix right now, and since it never made it into the main branch I don't think we should have any backwards compatibility issues with app controllers that are not part of cogent.