Updates Cogent to use Makeblastdb instead of the deprecated formatdb
Using the `assign_taxonomy.py` script with `-m blast` will result in the following error:

```cogent.app.util.ApplicationNotFoundError: Cannot find formatdb. Is it installed? Is it in your path?```

the problem here is that formatdb has been deprecated in favor of `makeblastdb` and it is called form this script. I changed the ClassName and references to the class but since only the functions from this file seem to be called directly this change should work fine as a fix. 

This is a quick first-pass at addressing: https://github.com/qiime/qiime/issues/1225